engines/requirements.txt
# Engines Service Dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=2.0.0
httpx>=0.24.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=1.5.0
psutil>=5.9.0
aiofiles>=23.0.0
asyncio-throttle>=1.0.0
